fix: pago ya no se hereda entre escaneos de Recepcion; tarifas enteras
Dos correcciones puntuales, ambas por regla de negocio explicita del
dueño del producto (no por refactor propio):

1. recepcion-client.tsx: "pago" (Pagado/Por pagar + monto) ya NO se
   mantiene entre escaneos. Antes, deliberadamente, se conservaba junto
   a cliente/recoge para no reescribirlo en varios paquetes del mismo
   remitente — pero eso significaba que si el operador olvidaba
   resetearlo antes de escanear el paquete de OTRO cliente que no pago,
   ese paquete quedaba marcado PAGADO sin que nadie hubiera pagado.
   Ahora registrar() resetea "pago" a PAGO_VACIO (Por pagar, tarifa
   base) apenas se acepta cada escaneo — mismo momento en que ya se
   limpiaba el campo de codigo, antes de esperar la respuesta del
   servidor. Si el mismo cliente trae varios paquetes pagados, el
   operador vuelve a marcar "Pagado" en cada uno.

2. tarifas/route.ts + series/route.ts + series/[inicial]/route.ts:
   tarifaBase, costoAdicionalDia y tarifaBaseOverride ahora exigen
   enteros (.int()) en la validacion. La especificacion es explicita en
   que el costo nunca genera centavos ni se prorratea, pero nada impedia
   antes que un ADMIN configurara ej. Bs2.50 y esos centavos se filtraran
   a cada paquete calculado por pricing.ts. Sin cambios a datos
   existentes (los valores configurados actualmente ya eran enteros).

TypeScript, ESLint, Prisma validate y build: todos limpios.

+++ src/app/api/configuracion/series/[inicial]/route.ts
@@ -7,7 +7,9 @@ import { registrarAuditoria, extraerContextoRequest } from '@/lib/auditoria';
 
 const bodySchema = z.object({
   descripcion: z.string().trim().min(1, 'La descripción es requerida').max(120),
-  tarifaBaseOverride: z.number().min(0).max(1_000_000).nullable().optional(),
+  // Entero, mismo criterio que tarifaBase en tarifas/route.ts: el costo
+  // nunca debe generar centavos.
+  tarifaBaseOverride: z.number().int().min(0).max(1_000_000).nullable().optional(),
   activo: z.boolean(),
 });
